requirements.txt
yfinance
pandas
numpy
gitpython
